fix: always isolate modules
Previous to this fix, if `dir` was passed as a string instead of a path
literal, then the isolation mechanisms provided by the implicit copy
behavior of a path in string interpolation was broken, and modules could
see the files in their directory.

Now both cases are treated identically and isolation is preserved.
`strToPath` takes a path like string and converts it back to a path
literal to accomplish this.