fix: work inside a flake
While we don't want to encourage it, it would be helpful to ease
adoption if users could call this module function from inside their
flake.nix. The only thing preventing that was the path referencing
becoming a pure /nix/store path reference implicitly (via the flake)
and so failing because of the string context.

Fix this by simply removing the string context. This isn't actually
unsafe since if the absolute path literal refers to a store path and is
ever coerced back to a string, nix will readd the context automatically.