chore: refactor for readability & tracability
The default.nix logic should now be easier to follow with core concerns
factored out into a seperate `src` directory. Also, a better error is
reported to the user in the event a module is not in a valid format,
showing the exact location on disk where the failed module resides.