feat: add internal features for the composer
This is the first implementation of the kinds of features atoms will
be able to set on themselves. The internal composer function has the
ability to define it's own features as well.

the semantic for declaring features is basically identical to doing so
in a Cargo.toml from Rust, for those famililar:
```toml
[features]
my_feature = []
my_other_feature = ["my_feature"]
default = ["my_feature"]
```

Here, the default feature enabled is `my_feature` because it is in the
`default` list. When `my_other_feature` is enabled, it also implicitly
enables `my_feature` since it is in it's list.