fix: `strToPath` shouldn't disgard context
Instead, we just need to prepare the dir at the beginning to ensure that
if it is a store path, we strip the context just once, which is safe to
do for our purpose, since when the path is imported to the store, the
string context will be readded.

This is inappropriate to do in a general library function though, so
we move it to it's own internal function.