fix: cleanup & refactor import code
The imported toml config is now loaded into `__atom` in the top level
scope, if when the user needs a ref. Perhaps this isn't the appropriate
place long term, but its a good place-holder that avoids polluting the
global public interface.

Also, change `[lib]` to `[atom]` to be more explicit, and fix the tests
to use toml files to declare their respective atoms. Also fixed an issue
with feature parsing.