fix: name should be set explicitly
In order to exhibit the same behavior in or out of `pure-eval`, we
need to use `builtins.path` and set the `name` attribute manually.

If we just rely on the implicit conversion of paths, then the name
will be based on a nix store path, carry more context than appropriate.

Also use `path.make` to hopefully make the semantic a little more
clear to the reader.