fix: allow npins sources to refer to each other
In order to pass other legacy inputs declared by `[fetch]` as arguments
to `import`, assume that any argument in the args list which is a set
and has a name which matches the name of an input should be substituted
for that input.

That is, in:
```toml
[fetch.pkgs]
name = "nixpkgs"
import = true
args = [{}]

[fetch.fenix]
import = true
args = [{pkgs = ""}]
```

The value of `pkgs` (`""`) is ignored in the toml, and is substituted
for the final value of `fetch.pkgs`.