feat: trivial debug tracing information
One of the best things about having a predictable module structure is
that we can trivially construct error information. We can then use this
to report unambiguously where we are in an Atom during evaluation.

This makes debugging trivial, since if you fail anywhere in your code
the debug trace (when enabled) will print the last location visited
inside the module system before the error occurred.