fix: soften the api a bit
aborting on builtin access is gonna cause more pain than required.

Instead we will simply make it a reference to `std`. In addition some
of the builtins have been downgraded from aborting to warnings which
return constants, if using a constant makes sense in that instance.

This should make migrating to atom a little more friendly and keep
us focused on those impurities which will actually cause us more pain
than necessary (e.g. `import`).

In order to make this work nicely we had to create a formal priority
ordering for keys that exist in both builtins and std.

Essentially `std` takes priority, but pure builtins will always still
be accesible via their __* variants in such a case. In cases where std
outputs a set named the same as a builtin, such as builtins.path,
then we take standards version, but apply a functor so that we can still
call the builtin function. e.g. `std.path.make` and
`std.path { name = "foo";}` are both valid, the former being a `std`
function, the latter a builtin.